Add Check to only remove thread creation messages in specific channels (ex. public ideas)

+++ dozer/cogs/hacks.py
@@ -22,7 +22,7 @@ MEDIA_CHANNEL_ID = 676583549561995274
 PUBLIC_CHANNEL_IDS = [320719178132881408, 676583549561995274, 771188718198456321]
 # media, robot-showcase
 EMBED_ONLY_CHANNEL_IDS = [676583549561995274, 771188718198456321]
-THREAD_REMOVE_CHANNEL_IDS = []
+THREAD_MESSAGE_REMOVE_CHANNEL_IDS = [674081079761829898]
 
 class Hacks(Cog):
 
@@ -113,6 +113,8 @@ _Please set your nickname with `%nick NAME - TEAM#` in #bot-spam to reflect your
 
     @Cog.listener
     async def on_thread_create(self, thread):
+        if (thread.parent.id is not in THREAD_MESSAGE_REMOVE_CHANNEL_IDS):
+            return
         async for message in thread.parent.history(limit=10):
             if (message.type == MessageType.thread_created and message.thread and message.thread.id == thread.id):
                 await message.delete()
